requirements.txt
setuptools==59.5.0
wheel==0.37.1
Flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
Pillow==9.3.0
numpy==1.21.0
https://download.pytorch.org/whl/cpu/torch-1.9.0%2Bcpu-cp39-cp39-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.10.0%2Bcpu-cp39-cp39-linux_x86_64.whl
timm==0.4.12
requests==2.28.1
python-dotenv==0.19.0
